set batch checkpoint date to correct date
The current code naively sets the checkpoint to the
datetime that the batch was downloaded as opposed to the
'since date' in the last item of the batch.

This means that any interruptions in the export
would not correctly restart from where they left off
but would instead only export data that was updated
since the last run.